feat: split Item-related models into new app
