+++ openedx_learning/apps/core/publish/models.py
@@ -10,9 +10,9 @@ associated with a LearningContext. So in that sense:
 
 """
 from django.db import models
+from django.conf import settings
 
 from openedx_learning.lib.fields import (
-    hash_field,
     identifier_field,
     immutable_uuid_field,
     manual_date_time_field,
@@ -41,19 +41,22 @@ class LearningContextVersion(models.Model):
     """
     uuid = immutable_uuid_field()
     learning_context = models.ForeignKey(LearningContext, on_delete=models.CASCADE)
+    prev_version = models.ForeignKey(LearningContext, on_delete=models.RESTRICT, null=True, related_name='+')
     created = manual_date_time_field()
 
 
+# Placeholder:
+"""
 class PublishLog(models.Model):
-    """
-    
-    """
     learning_context = models.ForeignKey(LearningContext, on_delete=models.CASCADE)
     version_num = models.PositiveIntegerField()
 
     # Note: The same LearningContextVersion can show up multiple times if it's
     # the case that something was reverted to an earlier version.
     learning_context_version = models.ForeignKey(LearningContextVersion, on_delete=models.RESTRICT)
+
+    published_at = manual_date_time_field()
+    published_by = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.SET_NULL, null=True)
 
     class Meta:
         constraints = [
@@ -69,104 +72,4 @@ class PublishLog(models.Model):
                 name="learning_publish_pl_uniq_lc_vn",
             )
         ]
-
-
-class ItemRaw(models.Model):
-    """
-    This is the most basic piece of raw Item content data, with no metadata.
-
-    ItemRaw stores data in a Binary BLOB `data` field. This data is not
-    normalized in any way, meaning that pieces of content that are semantically
-    equivalent (e.g. differently spaced/sorted JSON) will result in new entries.
-    This model is intentionally ignorant of what these things mean, because it
-    expects supplemental data models to build on top of it.
-
-    The other fields on ItemRaw are for data that is intrinsic to the file data
-    itself (e.g. the size). Any smart parsing of the contents into more
-    structured metadata should happen in other models that hang off of ItemInfo.
- 
-    ItemRaw models are not versioned in any way. The concept of versioning
-    exists at a higher level.
-
-    TODO: Size limit thoughts–configurable? Start at 10 MB?
-    """
-    learning_context = models.ForeignKey(LearningContext, on_delete=models.CASCADE)
-    hash_digest = hash_field()
-
-    # Per RFC 4288, MIME type and sub-type may each be 127 chars. Add one more
-    # char for the '/' in the middle, and we're at 255.
-    mime_type = models.CharField(max_length=255, blank=False, null=False)
-    size = models.PositiveBigIntegerField()
-
-    # This should be manually set so that multiple ItemRaw rows being set in the
-    # same transaction are created with the same timestamp. The timestamp should
-    # be UTC.
-    created = manual_date_time_field()
-
-    data = models.BinaryField(null=False)
-
-    class Meta:
-        constraints = [
-            # Make sure we don't store duplicates of this raw data within the
-            # same LearningContext, unless they're of different mime types.
-            models.UniqueConstraint(
-                fields=["learning_context_id", "mime_type", "hash_digest"],
-                name="learning_publishing_item_raw_uniq_lc_hd",
-            )
-        ]
-
-
-class ItemInfo(models.Model):
-    """
-
-    """
-    uuid = immutable_uuid_field()
-    learning_context = models.ForeignKey(LearningContext, on_delete=models.CASCADE)
-    type = models.CharField(max_length=100)
-    item_raw = models.ForeignKey(ItemRaw, on_delete=models.RESTRICT)
-
-    created = manual_date_time_field()
-
-
-class Item(models.Model):
-    """
-    This represents any content that has ever existed in a LearningContext.
-    """
-    uuid = immutable_uuid_field()
-    learning_context = models.ForeignKey(LearningContext, on_delete=models.CASCADE)
-    identifier = identifier_field()
-
-    published_version = models.ForeignKey('ItemVersion', on_delete=models.RESTRICT, null=True, related_name="+")
-    first_published_at = models.DateTimeField(null=True)  # can be null if never published 
-    last_published_at = models.DateTimeField(null=True)   # can be null if never published
-
-    created = manual_date_time_field()
-    modified = manual_date_time_field()
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["learning_context_id", "identifier"],
-                name="learning_publishing_lcb_one_identifier_per_lc",
-            )
-        ]
-
-    def __str__(self):
-        return f"{self.identifier}"
-
-
-class ItemVersion(models.Model):
-    """
-    """
-    uuid = immutable_uuid_field()
-    item_info = models.ForeignKey(ItemInfo, on_delete=models.RESTRICT)
-    item = models.ForeignKey(Item, on_delete=models.CASCADE)
-    title = models.CharField(max_length=1000, blank=True, null=True)
-
-    def __str__(self):
-        return f"{self.uuid}: {self.title}"
-
-
-class LearningContextVersionItemVersion(models.Model):
-    learning_context_version = models.ForeignKey(LearningContextVersion, on_delete=models.CASCADE)
-    item_version = models.ForeignKey(ItemVersion, on_delete=models.PROTECT)
+"""
